+++ e2e-tests/commands/insurance/complete-prepare-application-section-multiple-policy-type.js
@@ -46,7 +46,7 @@ const completePrepareApplicationMultiplePolicyType = ({
 
   cy.completeAndSubmitNatureOfYourBusiness();
   cy.completeAndSubmitTurnoverForm();
-  cy.completeAndSubmitCreditControlForm();
+  cy.completeAndSubmitCreditControlForm({});
   cy.completeAndSubmitBrokerForm({ usingBroker });
 
   submitButton().click();