+++ e2e-tests/constants/field-ids/insurance/business/index.js
@@ -19,7 +19,7 @@ export const EXPORTER_BUSINESS = {
     ESTIMATED_ANNUAL_TURNOVER: 'estimatedAnnualTurnover',
     PERCENTAGE_TURNOVER: 'exportsTurnoverPercentage',
   },
-  CREDIT_CONTROL: 'creditControl',
+  HAS_CREDIT_CONTROL: 'hasCreditControlProcess',
   BROKER: {
     LEGEND: 'broker',
     USING_BROKER: 'isUsingBroker',