+++ e2e-tests/insurance/cypress/e2e/journeys/your-business/change-your-answers/change-your-answers-change-broker.spec.js
@@ -39,7 +39,7 @@ context('Insurance - Your business - Change your answers - Broker - As an export
       cy.completeAndSubmitCompanyDetails({});
       cy.completeAndSubmitNatureOfYourBusiness();
       cy.completeAndSubmitTurnoverForm();
-      cy.completeAndSubmitCreditControlForm();
+      cy.completeAndSubmitCreditControlForm({});
       cy.completeAndSubmitBrokerForm({ usingBroker: true });
 
       url = `${Cypress.config('baseUrl')}${ROOT}/${referenceNumber}${CHECK_YOUR_ANSWERS}`;