+++ src/ui/server/controllers/insurance/business/credit-control/index.ts
@@ -6,6 +6,7 @@ import singleInputPageVariables from '../../../../helpers/page-variables/single-
 import getUserNameFromSession from '../../../../helpers/get-user-name-from-session';
 import constructPayload from '../../../../helpers/construct-payload';
 import generateValidationErrors from '../../../../shared-validation/yes-no-radios-form';
+import mapAndSave from '../map-and-save/business';
 import { Request, Response } from '../../../../../types';
 
 const {
@@ -14,9 +15,9 @@ const {
   PROBLEM_WITH_SERVICE,
 } = INSURANCE_ROUTES;
 
-const { CREDIT_CONTROL } = BUSINESS_FIELD_IDS;
+const { HAS_CREDIT_CONTROL } = BUSINESS_FIELD_IDS;
 
-export const FIELD_ID = CREDIT_CONTROL;
+export const FIELD_ID = HAS_CREDIT_CONTROL;
 
 export const TEMPLATE = TEMPLATES.INSURANCE.EXPORTER_BUSINESS.CREDIT_CONTROL;
 
@@ -47,6 +48,7 @@ export const get = (req: Request, res: Response) => {
         BACK_LINK: req.headers.referer,
       }),
       userName: getUserNameFromSession(req.session.user),
+      application: application.business,
     });
   } catch (err) {
     console.error('Error getting credit control %O', err);
@@ -61,7 +63,7 @@ export const get = (req: Request, res: Response) => {
  * @param {Express.Response} Express response
  * @returns {Express.Response.redirect} Next part of the flow or error page
  */
-export const post = (req: Request, res: Response) => {
+export const post = async (req: Request, res: Response) => {
   try {
     const { application } = res.locals;
 
@@ -85,6 +87,16 @@ export const post = (req: Request, res: Response) => {
         validationErrors,
         submittedValues: payload,
       });
+    }
+
+    /**
+     * No validation errors.
+     * call mapAndSave to call the API.
+     */
+    const saveResponse = await mapAndSave.business(payload, application);
+
+    if (!saveResponse) {
+      return res.redirect(PROBLEM_WITH_SERVICE);
     }
 
     return res.redirect(`${INSURANCE_ROOT}/${referenceNumber}${BROKER_ROOT}`);